chore: PostResponseDto count 타입 수정
#14

+++ src/main/java/com/stemm/pubsub/service/post/dto/response/PostResponseDto.java
@@ -9,7 +9,7 @@ public record PostResponseDto(
     String imageUrl,
     LocalDateTime createdDate,
     LocalDateTime lastModifiedDate,
-    long likeCount,
-    long dislikeCount
+    int likeCount,
+    int dislikeCount
 ) {
 }
